Version 0.7.0
Changes:
- add normatives mocks
- split territories endpoints code to multiple files
