--- pyproject copy.toml
@@ -1,62 +1,0 @@
-[tool.poetry]
-name = "urban-api"
-version = "0.1.0"
-description = "This is a Digital Territories Platform API to access and manipulate basic territories data."
-authors = ["Babayev Ruslan <rus.babaef@yandex.ru>"]
-readme = "README.md"
-
-[tool.poetry.dependencies]
-python = "^3.10"
-fastapi = "^0.110.2"
-asyncio = "^3.4.3"
-asyncpg = "^0.29.0"
-uvicorn = "^0.29.0"
-python-keycloak = "^3.12.0"
-pydantic = "^2.7.1"
-alembic = "^1.13.1"
-loguru = "^0.7.2"
-sqlalchemy = { version = "^2.0.29", extras = ["asyncio"] }
-cachetools = "^5.3.3"
-geoalchemy2 = "^0.15.0"
-shapely = "^2.0.4"
-pandas = "^2.2.2"
-exceptiongroup = "^1.2.1"
-
-[tool.poetry.scripts]
-launch_urban_api = "urban_api.__main__:main"
-
-[tool.poetry.group.dev.dependencies]
-black = "^24.4.2"
-pylint = "^3.1.0"
-isort = "^5.13.2"
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
-
-[tool.black]
-line-length = 120
-target-version = ['py310']
-
-[tool.pylint.format]
-max-line-length = 120
-expected-line-ending-format = "LF"
-max-locals = 20
-max-attributes = 10
-max-args = 9
-extension-pkg-allow-list = ["networkit"]
-disable = [
-    "duplicate-code",
-    "missing-module-docstring",
-    "missing-class-docstring",
-    "missing-function-docstring",
-]
-
-[tool.isort]
-multi_line_output = 3
-include_trailing_comma = true
-force_grid_wrap = 0
-use_parentheses = true
-ensure_newline_before_comments = true
-line_length = 120
-split_on_trailing_comma = true
