+++ urban_api/endpoints/v1/routers.py
@@ -6,9 +6,9 @@ It is needed to import files which use these routers to initialize endpoints.
 
 from fastapi import APIRouter
 
-indicators_router = APIRouter(tags=["indicators"], prefix="/v1")
+from .territories import routers_list as territories_routers
 
-territories_router = APIRouter(tags=["territories"], prefix="/v1")
+indicators_router = APIRouter(tags=["indicators"], prefix="/v1")
 
 service_types_router = APIRouter(tags=["service_types"], prefix="/v1")
 
@@ -22,12 +22,12 @@ services_router = APIRouter(tags=["services"], prefix="/v1")
 
 routers_list = [
     indicators_router,
-    territories_router,
     services_router,
     functional_zones_router,
     physical_objects_router,
     object_geometries_router,
     service_types_router,
+    *territories_routers,
 ]
 
 __all__ = [
