+++ urban_api/endpoints/v1/territories/services.py
@@ -1,0 +1,92 @@
+"""Services territories-related endpoints are defined here."""
+
+from fastapi import Depends, Path, Query
+from fastapi_pagination import paginate
+from sqlalchemy.ext.asyncio import AsyncConnection
+from starlette import status
+
+from urban_api.db.connection import get_connection
+from urban_api.logic.territories import (
+    get_services_by_territory_id_from_db,
+    get_services_capacity_by_territory_id_from_db,
+    get_services_with_geometry_by_territory_id_from_db,
+)
+from urban_api.schemas import ServicesData, ServicesDataWithGeometry
+from urban_api.schemas.pages import Page
+
+from .routers import territories_router
+
+
+@territories_router.get(
+    "/territory/{territory_id}/services",
+    response_model=Page[ServicesData],
+    status_code=status.HTTP_200_OK,
+)
+async def get_services_by_territory_id(
+    territory_id: int = Path(description="territory id", gt=0),
+    service_type_id: int | None = Query(None, description="Service type id", gt=0),
+    name: str | None = Query(None, description="Filter services by name substring (case-insensitive)"),
+    connection: AsyncConnection = Depends(get_connection),
+) -> Page[ServicesData]:
+    """
+    Summary:
+        Get services for territory
+
+    Description:
+        Get services for territory by id, service type and name could be specified in parameters
+    """
+
+    services = await get_services_by_territory_id_from_db(territory_id, connection, service_type_id, name)
+    services = [ServicesData.from_dto(service) for service in services]
+
+    return paginate(services)
+
+
+@territories_router.get(
+    "/territory/{territory_id}/services_with_geometry",
+    response_model=Page[ServicesDataWithGeometry],
+    status_code=status.HTTP_200_OK,
+)
+async def get_services_with_geometry_by_territory_id(
+    territory_id: int = Path(description="territory id", gt=0),
+    service_type_id: int | None = Query(None, description="Service type id", gt=0),
+    name: str | None = Query(None, description="Filter services by name substring (case-insensitive)"),
+    connection: AsyncConnection = Depends(get_connection),
+) -> Page[ServicesDataWithGeometry]:
+    """
+    Summary:
+        Get services for territory
+
+    Description:
+        Get services for territory by id, service type could be specified in parameters
+    """
+
+    services = await get_services_with_geometry_by_territory_id_from_db(territory_id, connection, service_type_id, name)
+    services = [ServicesDataWithGeometry.from_dto(service) for service in services]
+
+    return paginate(services)
+
+
+@territories_router.get(
+    "/territory/{territory_id}/services_capacity",
+    response_model=int,
+    status_code=status.HTTP_200_OK,
+)
+async def get_total_services_capacity_by_territory_id(
+    territory_id: int = Path(description="territory id", gt=0),
+    service_type_id: int | None = Query(None, description="Service type id", gt=0),
+    connection: AsyncConnection = Depends(get_connection),
+) -> int:
+    """
+    Summary:
+        Get services capacity for territory
+
+    Description:
+        Get aggregated capacity of services for territory
+    """
+
+    capacity = await get_services_capacity_by_territory_id_from_db(
+        territory_id, connection, service_type_id=service_type_id
+    )
+
+    return capacity
