+++ urban_api/version.py
@@ -1,4 +1,4 @@
 """Version constants are defined here and should be updated on release."""
 
-VERSION = "0.6.1"
-LAST_UPDATE = "2024-06-19"
+VERSION = "0.7.0"
+LAST_UPDATE = "2024-06-20"
